+++ src/Components/Inputs/index.tsx
@@ -1,4 +1,6 @@
 export * from './Input'
+export * from './InputLabel'
 export * from './TextArea'
+export * from './Toggle'
 export * from './Composed/AutoInput'
 export * from './Composed/RangeSlider'